Crate massa_db_exports

Modules

constants 🔒

controller 🔒

db_batch 🔒

error 🔒

Copyright (c) 2022 MASSA LABS info@massa.net This file defines all error types for final state management

settings 🔒

Structs

MassaDBConfig

Config structure for a MassaDBRaw

StreamBatch

A Batch of elements from the database, used by a bootstrap server.

Enums

MassaDBError

Massa DB error

MassaDirection

Similar to RocksDB’s Direction

MassaIteratorMode

Similar to RocksDB’s IteratorMode

Constants

ASYNC_POOL_PREFIX

CF_ERROR

CHANGE_ID_DESER_ERROR

CHANGE_ID_KEY

CHANGE_ID_SER_ERROR

CRUD_ERROR

CYCLE_HISTORY_DESER_ERROR

CYCLE_HISTORY_PREFIX

CYCLE_HISTORY_SER_ERROR

DEFERRED_CREDITS_DESER_ERROR

DEFERRED_CREDITS_PREFIX

DEFERRED_CREDITS_SER_ERROR

EXECUTED_DENUNCIATIONS_INDEX_DESER_ERROR

EXECUTED_DENUNCIATIONS_INDEX_SER_ERROR

EXECUTED_DENUNCIATIONS_PREFIX

EXECUTED_OPS_ID_DESER_ERROR

EXECUTED_OPS_ID_SER_ERROR

EXECUTED_OPS_PREFIX

EXECUTION_TRAIL_HASH_PREFIX

KEY_DESER_ERROR

KEY_LEN_SER_ERROR

KEY_SER_ERROR

LEDGER_PREFIX

MESSAGE_DESER_ERROR

MESSAGE_ID_DESER_ERROR

MESSAGE_ID_SER_ERROR

MESSAGE_SER_ERROR

METADATA_CF

MIP_STORE_PREFIX

MIP_STORE_STATS_PREFIX

OPEN_ERROR

STATE_CF

STATE_HASH_BYTES_LEN

STATE_HASH_ERROR

STATE_HASH_INITIAL_BYTES

STATE_HASH_KEY

VERSIONING_CF

Traits

MassaDBController

Controller trait for the MassaDB TODO: MOCK IT WITH MOCKALL. HAVING LIFETIMES ERRORS WITH AUTO MOCK

Type Aliases

DBBatch

We use batching to reduce the number of writes to the database

Key

ShareableMassaDBController

Value